+++ config/telescope.php
@@ -76,6 +76,7 @@ return [
     */
 
     'middleware' => [
+        'web',
         'auth',
         Authorize::class,
     ],