+++ resources/lang/ar/dashboard.php
@@ -13,7 +13,7 @@ return [
     |
     */
 
-    'Settings' => 'إعدادات',
+    'Settings' => 'الإعدادات',
     'My profile' => 'صفحتي',
     'Activity' => 'أنشطة',
     'Support' => 'الدعم',
@@ -27,10 +27,23 @@ return [
     'Maps' => 'خرائط',
     'Login' => 'تسجيل الدخول',
     'Register' => 'إنشاء حساب',
-    '' => '',
-    '' => '',
-    '' => '',
-    '' => '',
-    
-
+    'Anime Settings' => 'إعدادات الأنمي',
+    'Episodes Settings' => 'إعدادات الحلقات',
+    'Blog Settings' => 'إعدادات المدونة',
+    'Comments Settings' => 'إعدادات التعليقات',
+    'Frontend Settings' => 'إعدادات الواجهة',
+    'Social Settings' => 'إعدادات مواقع التواصل',
+    'Backup Settings' => 'إعدادات النسخ الإحتياطي',
+    'Users Settings' => 'إعدادات المستخدمين',
+    'Advanced Settings' => 'إعدادات متقدمة',
+    'Sitemap Settings' => 'إعدادات خرائط الموقع',
+    'Update All Sitemaps' => 'تحديث كل خرائط الموقع',
+    'Update Animes Sitemaps' => 'تحديث خرائط الأنميات',
+    'Update Episodes Sitemaps' => 'تحديث كل خرائط الحلقات',
+    'Update Pages Sitemaps' => 'تحديث كل خرائط الصفحات',
+    'Update Blog Sitemaps' => 'تحديث كل خرائط المدونة',
+    'Update Main Sitemap' => 'تحديث خريطة الموقع الرئيسية',
+    'Settings Updated successfully' => 'تم تحديث كل الإعدادات بنجاح',
+    'Save' => 'حفظ',
+    'Update' => 'تحديث'
 ];