+++ routes/web.php
@@ -1,33 +1,23 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Http;
 
 // the home page of the project
-Route::middleware('page-cache')->get('/', 'WelcomeController@index');
+Route::get('/', 'WelcomeController@index');
+Route::get('/test', 'WelcomeController@test');
 
-// Route::get('/links/{id}', 'LinkController@index');
-// Route::post('/links/{anime}', 'LinkController@store');
+Route::get('/req', function () {
+    $response = Http::asForm()->post('https://www.animesilver.com/watch/getQAServer', [
+        'auth' => 'getQAServer',
+        'ep' => 'الحلقة 06',
+        'id' => '2242',
+        'server' => 'Premieum',
+        'c' => '1',
+    ]);
+    return $response;
+});
 
-Route::domain('www.royalanime.com')->group(function () {
-    Route::get('domain-test/{id}', function ($id) {
-        return $id . ' from www';
-    });
-});
-Route::domain('cdn.royalanime.com')->group(function () {
-    Route::get('domain-test/{id}', function ($id) {
-        return $id . 'from cdn';
-    });
-});
-// Route::get('/req', function () {
-//     $response = Http::asForm()->post('https://www.animesilver.com/watch/getQAServer', [
-//         'auth' => 'Premieum',
-//         'ep' => 'الحلقة 05',
-//         'id' => '2255',
-//         'server' => 'Oserver',
-//         'c' => '1',
-//     ]);
-//     return $response;
-// });
     
 Route::resources([
     'animes.episodes' => 'EpisodeController',   // All Episodes of an Anime. This route has another route for slug
@@ -45,6 +35,7 @@ Route::resources([
  */
 Route::get('animes/{anime}/{slug?}', ['as'=>'animes.show', 'uses'=>'AnimeController@show']);
 Route::get('animes/{anime}/episodes/{episode}/{slug?}', ['as'=>'animes.episodes.show', 'uses'=>'EpisodeController@show']);
+Route::get('download_links', ['as'=>'animes.episodes.show', 'uses'=>'EpisodeController@download_links']);
 Route::get('tags/{tag}/{slug?}', ['as'=>'tags.show', 'uses'=>'TagController@show']);
 Route::get('blog/posts/{post}/{slug?}', ['as'=>'posts.show', 'uses'=>'PostController@show']);
     
@@ -73,15 +64,34 @@ Route::group(['middleware' => 'auth'], function () {
     Route::put('profile/password', ['as' => 'profile.password', 'uses' => 'ProfileController@password']);
 
     // Settings Pages
-    Route::get('settings',              ['as' => 'settings.index',      'uses' => 'SettingController@index']);
-    Route::get('settings/animes',       ['as' => 'settings.animes',     'uses' => 'SettingController@animes']);
-    Route::get('settings/episodes',     ['as' => 'settings.episodes',   'uses' => 'SettingController@episodes']);
-    Route::get('settings/blog',         ['as' => 'settings.blog',       'uses' => 'SettingController@blog']);
-    Route::get('settings/comments',     ['as' => 'settings.comments',   'uses' => 'SettingController@comments']);
-    Route::get('settings/frontend',     ['as' => 'settings.frontend',   'uses' => 'SettingController@frontend']);
-    Route::get('settings/social',       ['as' => 'settings.social',     'uses' => 'SettingController@social']);
-    Route::get('settings/backup',       ['as' => 'settings.backup',     'uses' => 'SettingController@backup']);
-    Route::get('settings/users',        ['as' => 'settings.users',      'uses' => 'SettingController@users']);
+    Route::get('settings',              ['as' => 'settings.index',              'uses' => 'SettingController@index']);
+    Route::get('settings/animes',       ['as' => 'settings.animes',             'uses' => 'SettingController@animes']);
+    Route::put('settings/animes',       ['as' => 'settings.animes.update',      'uses' => 'SettingController@update_animes']);
+
+    Route::get('settings/episodes',     ['as' => 'settings.episodes',           'uses' => 'SettingController@episodes']);
+    Route::put('settings/episodes',     ['as' => 'settings.episodes.update',    'uses' => 'SettingController@update_episodes']);
+
+    Route::get('settings/blog',         ['as' => 'settings.blog',               'uses' => 'SettingController@blog']);
+    Route::put('settings/blog',         ['as' => 'settings.blog.update',        'uses' => 'SettingController@update_blog']);
+
+    Route::get('settings/comments',     ['as' => 'settings.comments',           'uses' => 'SettingController@comments']);
+    Route::put('settings/comments',     ['as' => 'settings.comments.update',    'uses' => 'SettingController@update_comments']);
+
+    Route::get('settings/frontend',     ['as' => 'settings.frontend',           'uses' => 'SettingController@frontend']);
+    Route::put('settings/frontend',     ['as' => 'settings.frontend.update',    'uses' => 'SettingController@update_frontend']);
+
+    Route::get('settings/social',       ['as' => 'settings.social',             'uses' => 'SettingController@social']);
+    Route::put('settings/social',       ['as' => 'settings.social.update',      'uses' => 'SettingController@update_social']);
+
+    Route::get('settings/backup',       ['as' => 'settings.backup',             'uses' => 'SettingController@backup']);
+    Route::put('settings/backup',       ['as' => 'settings.backup.update',      'uses' => 'SettingController@update_backup']);
+
+    Route::get('settings/users',        ['as' => 'settings.users',              'uses' => 'SettingController@users']);
+    Route::put('settings/users',        ['as' => 'settings.users.update',       'uses' => 'SettingController@update_users']);
+
+    Route::get('settings/advanced',     ['as' => 'settings.advanced',           'uses' => 'SettingController@advanced']);
+    Route::put('settings/advanced',     ['as' => 'settings.advanced.update',    'uses' => 'SettingController@update_advanced']);
+
 });